Reorganize all files and data
