+++ lib/modules/league_stats_module.rb
@@ -1,49 +1,6 @@
 require 'pry'
 module LeagueStats
 
-  def convert_id_to_name(id)
-    team = @teams.find {|team| team.team_id == id}
-    team.team_name
-  end
-
-  def home_team_goals
-    goals_home_team = Hash.new(0)
-    @games.each do |game|
-      goals_home_team[game.home_team_id] += game.home_goals
-    end
-    goals_home_team
-  end
-
-  def away_team_goals
-    goals_away_team = Hash.new(0)
-    @games.each do |game|
-      goals_away_team[game.away_team_id] += game.away_goals
-    end
-    goals_away_team
-  end
-
-  def home_games_played
-    games_home_played = Hash.new(0)
-    @games.each do |game|
-      games_home_played[game.home_team_id] += 1
-    end
-    games_home_played
-  end
-
-  def away_games_played
-    games_away_played = Hash.new(0)
-    @games.each do |game|
-      games_away_played[game.away_team_id] += 1
-    end
-    games_away_played
-  end
-
-  def total_games_played
-    total_games_played = away_games_played.merge!(home_games_played) do |id, ag, hg|
-      ag + hg
-    end
-  end
-
   def highest_scoring_home_team
     high_team = home_team_goals.max_by do |team, goals|
           (goals.to_f / home_games_played[team])
@@ -76,31 +33,6 @@ module LeagueStats
     @teams.count {|team| team.team_id}
   end
 
-
-  def total_goals_by_game_team
-    total_goal_by_team = Hash.new(0)
-    @game_teams.each do |game|
-      total_goal_by_team[game.team_id] += game.goals.to_f
-    end
-    total_goal_by_team
-  end
-
-  def total_games_by_game_team
-   total_games_by_teams = Hash.new(0)
-   @game_teams.each do |game|
-      total_games_by_teams[game.team_id] += 1
-    end
-    total_games_by_teams
-  end
-
-  def average_offense
-    avg_offense = Hash.new
-    total_goals_by_game_team.map do |team_id, total_goals|
-      avg_offense[team_id] = (total_goals_by_game_team[team_id]/total_games_by_game_team[team_id]).round(2)
-    end
-    avg_offense
-  end
-
   def best_offense
     avg_offense = average_offense
     best_offense_team = avg_offense.max_by {|team_id, avg_goals| avg_goals}
@@ -111,29 +43,6 @@ module LeagueStats
     avg_offense = average_offense
     best_offense_team = avg_offense.min_by {|team_id, avg_goals| avg_goals}
     convert_id_to_name(best_offense_team[0])
-  end
-
-  def home_team_goals_against
-    goals_against_home_team = Hash.new(0)
-    @games.each do |game|
-      goals_against_home_team[game.home_team_id] += game.away_goals
-    end
-    goals_against_home_team
-  end
-
-  def away_team_goals_against
-    goals_against_away_team = Hash.new(0)
-    @games.each do |game|
-      goals_against_away_team[game.away_team_id] += game.home_goals
-    end
-    goals_against_away_team
-  end
-
-  def total_goals_against
-    total_goals_against = away_team_goals_against.merge!(home_team_goals_against) do |id, aga, hga|
-      aga + hga
-    end
-    total_goals_against
   end
 
   def best_defense
@@ -150,16 +59,6 @@ module LeagueStats
     convert_id_to_name(lowest_defense[0])
   end
 
-  def games_won_game_team
-    game_team_wins = Hash.new(0)
-    @game_teams.each do |game|
-      if game.won == 'TRUE'
-        game_team_wins[game.team_id] += 1
-      end
-    end
-    game_team_wins
-  end
-
   def winningest_team
     awesomest_team = games_won_game_team.max_by do |team_id, games_won|
       games_won.to_f / total_games_by_game_team[team_id]
@@ -167,49 +66,9 @@ module LeagueStats
     convert_id_to_name(awesomest_team[0])
   end
 
-  def change_hash_ids_to_name(hash)
-   combined_names = {}
-   hash.each do |team_id, games_won|
-     combined_names[convert_id_to_name(games_won)] = games_won
-   end
-   combined_names
- end
-
   def percentage_home_wins
     game_wins = @games.find_all {|game| game.outcome.include?('home win')}
     (game_wins.length / @games.length.to_f).round(2)
-  end
-
-  def total_home_wins_by_team
-    home_wins = Hash.new(0)
-    @game_teams.each do |team|
-      if team.won && team.hoa == 'home'
-        home_wins[team.team_id] += 1
-      end
-    end
-    home_wins
-  end
-
-  def home_wins_percentage_by_team
-    win_percentages = total_home_wins_by_team.merge(total_games_by_game_team){|team, win, games| win/games.to_f}
-    max = win_percentages.max_by{|k,v| v}
-    max[1].round(2)
-  end
-
-  def total_away_wins_by_team
-    away_wins = Hash.new(0)
-    @game_teams.each do |team|
-      if team.won && team.hoa == 'away'
-        away_wins[team.team_id] += 1
-      end
-    end
-    away_wins
-  end
-
-  def away_wins_percentage_by_team
-    aw_percentages = total_away_wins_by_team.merge(total_games_by_game_team){|team, win, games| win/games.to_f}
-    max = aw_percentages.max_by{|k,v| v}
-    max[1].round(2)
   end
 
   def worst_fans
@@ -224,74 +83,8 @@ module LeagueStats
     end
   end
 
-  def home_game_team_wins
-    home_game_team_win = Hash.new(0)
-      @game_teams.find_all do |game|
-        if game.won == 'TRUE' && game.hoa == 'home'
-        home_game_team_win[game.team_id] += 1
-        end
-      end
-    home_game_team_win
-  end
-
-  def away_game_team_wins
-    away_game_team_wins = Hash.new(0)
-      @game_teams.find_all do |game|
-        if game.won == 'TRUE' && game.hoa == 'away'
-        away_game_team_wins[game.team_id] += 1
-        end
-      end
-    away_game_team_wins
-  end
-
-  def total_home_games_played
-    total_home_games_played = Hash.new(0)
-    @game_teams.each do |game|
-      if game.hoa == 'home'
-      total_home_games_played[game.team_id] += 1
-      end
-    end
-    total_home_games_played
-  end
-
-  def total_away_games_played
-    total_away_game_played = Hash.new(0)
-    @game_teams.each do |game|
-      if game.hoa == 'away'
-      total_away_game_played[game.team_id] += 1
-      end
-    end
-    total_away_game_played
-  end
-
-  def percent_of_home_games_won
-    percent_of_home_games_won = Hash.new(0)
-      home_game_team_wins.map do |team_id, home_wins|
-      percent_of_home_games_won[team_id] = home_wins/total_home_games_played[team_id].to_f
-    end
-    percent_of_home_games_won
-  end
-
-  def percent_of_away_games_won
-    percent_of_away_games_won = Hash.new(0)
-      away_game_team_wins.map do |team_id, away_wins|
-      percent_of_away_games_won[team_id] = away_wins/total_away_games_played[team_id].to_f
-    end
-    percent_of_away_games_won
-  end
-
-  def difference_home_vs_away_won
-    difference_home_vs_away_won = Hash.new(0)
-      percent_of_home_games_won.map do |team_id, home_win|
-      difference_home_vs_away_won[team_id] = home_win - percent_of_away_games_won[team_id]
-    end
-    difference_home_vs_away_won
-  end
-
   def best_fans
     team_diff_greatest_home_v_away = difference_home_vs_away_won.max_by {|team_id, diff_win| diff_win }
-    team_diff_greatest_home_v_away
     convert_id_to_name(team_diff_greatest_home_v_away[0])
   end
-
 end
