+++ lib/modules/team_stat_helper_module.rb
@@ -1,0 +1,3 @@
+module TeamStatHelpers
+
+end
