--- lib/modules/team_stats.rb
@@ -1,9 +1,0 @@
-module TeamStats
-
-
-  def best_season
-    @teams.max_by do |team|
-      (team.data[])
-  end
-
-end
