+++ lib/modules/team_stats_module.rb
@@ -1,0 +1,3 @@
+module TeamStats
+
+end
