+++ lib/stat_tracker.rb
@@ -5,10 +5,19 @@ require_relative './teams'
 require_relative './game_teams'
 require_relative './modules/game_stats_module'
 require_relative './modules/league_stats_module'
+require_relative './modules/league_stat_helper_module'
+require_relative './modules/team_stats_module'
+require_relative './modules/team_stat_helper_module'
+require_relative './modules/global_helper_module'
+
 
 class StatTracker
     include GameStats
+    include GlobalHelpers
     include LeagueStats
+    include LeagueStatHelpers
+    include TeamStats
+    include TeamStatHelpers
 
 attr_reader :games, :teams, :game_teams
