+++ test/test_helper.rb
@@ -6,6 +6,11 @@ require 'minitest/autorun'
 require 'minitest/pride'
 require './lib/modules/game_stats_module'
 require './lib/modules/league_stats_module'
+require './lib/modules/league_stat_helper_module'
+require './lib/modules/team_stats_module'
+require './lib/modules/team_stat_helper_module'
+require './lib/modules/global_helper_module'
+
 require './lib/teams'
 require './lib/game'
 require './lib/game_teams'
